+++ src/Packman/Defaults.cs
@@ -6,6 +6,6 @@
         public static string DefaultLocalPath { get; set; } = "lib";
 
         public static int CacheDays { get; set; } = 3;
-        public static string CachePath { get; set; } = "%userprofile%\\.librarian";
+        public static string CachePath { get; set; } = "%userprofile%\\.packman";
     }
 }